# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/:nome": {
            "get": {
                "description": "Rota que recebendo um nome, saúda o usuário com o nome fornecido",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "saudacao"
                ],
                "summary": "API faz uma saudação",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/alunos": {
            "get": {
                "description": "Rota para exibir todos os alunos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alunos"
                ],
                "summary": "Exibe todos os alunos",
                "responses": {
                    "200": {
                        "description": "Modelo do aluno",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                }
            },
            "post": {
                "description": "Rota para criar um novo aluno",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aluno"
                ],
                "summary": "Cria um novo aluno",
                "parameters": [
                    {
                        "description": "Modelo do aluno",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/alunos/:id": {
            "get": {
                "description": "Rota para buscar um aluno por id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aluno"
                ],
                "summary": "Busca um aluno por id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "Rota para deletar um aluno, por id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aluno"
                ],
                "summary": "Deleta um aluno",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                }
            },
            "patch": {
                "description": "Rota para editar um aluno, por id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aluno"
                ],
                "summary": "Edita um aluno",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/alunos/cpf/:cpf": {
            "get": {
                "description": "Rota para buscar um aluno por cpf",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aluno"
                ],
                "summary": "Busca um aluno por cpf",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {
                "type": "any"
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Aluno": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "rg": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}